Fix code that adjusts order total with fee
Code that adjusts order total results in \PrestaShop\Decimal\Number object
being stored inside $order->total_paid_tax_excl property.

This is definitely not an expected value.

Prestashop "survives" this by pure luck, as the value is converted to
string during sql update. That might not be the case in the future.

In thirtybees platform this leads to invalid value being stored inside
database table already.